[BUGFIX] Harden class loading in non composer mode
We now always add our class prefixes to the main
class loader and disable other registration for
console classes.

This helps to fix an edge case where old class loading
info has overwritten what was added before.

While doing so, all non composer class loading code
is extracted to a autoload.php file, which is only
loaded when console is not installed with composer.

With that, we can remove remaining class_exists and
one composer mode check from bootstrap.

Fixes: #667
