+++ Resources/Private/ExtensionArtifacts/ext_emconf.php
@@ -29,4 +29,6 @@ $EM_CONF[$_EXTKEY] = [
     'suggests' => [
     ],
   ],
+  'autoload' => [
+  ],
 ];
