+++ Resources/Private/ExtensionArtifacts/ext_localconf.php
@@ -3,6 +3,7 @@ defined('TYPO3_MODE') or die('Access denied.');
 
 (function () {
     if ((TYPO3_REQUESTTYPE & TYPO3_REQUESTTYPE_CLI) || (TYPO3_MODE === 'BE' && isset($_GET['M']) && 'tools_ExtensionmanagerExtensionmanager' === $_GET['M'])) {
+        require \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath('typo3_console') . 'Resources/Private/ExtensionArtifacts/autoload.php';
         $signalSlotDispatcher = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Extbase\SignalSlot\Dispatcher::class);
         $signalSlotDispatcher->connect(
             \TYPO3\CMS\Extensionmanager\Utility\InstallUtility::class,
